+++ src/main/kotlin/homeworks/homework4/task1/hashFunctions/interfaces/IHashFunction.kt
@@ -1,0 +1,5 @@
+package homeworks.homework4.task1.hashFunctions.interfaces
+
+interface IHashFunction {
+    fun getHash(value: String, tableSize: Int): Int
+}
